web_app/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-dotenv
faker
azure-eventhub
